requirements.txt
fastapi
uvicorn
# required for embeddings
openai
pinecone-client
pandas
matplotlib
plotly
scipy
scikit-learn